refactor(website): rename DateTimeInMonths.tsx to utcDates.ts
The module no longer only computes a date some months out — it also
exports utcToday — and it contains no JSX, so neither the name nor the
.tsx extension fitted. utcDates.ts matches the camelCase .ts convention
used by the other plain helpers in src/utils.

Also use the .ts extension consistently across all four import sites,
which previously mixed extension and extensionless specifiers.
